# ai-pm-mindmap/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI PM Mindmap API",
    "version": "1.0.0",
    "description": "API powering the AI PM Mindmap experience. Provides Merge Request, Story, Test, and tree operations."
  },
  "paths": {
    "/api/merge-requests": {
      "get": {
        "summary": "List merge requests",
        "responses": {
          "200": {
            "description": "List of merge requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MergeRequest"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create merge request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created merge request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeRequest"
                }
              }
            }
          }
        }
      }
    },
    "/api/merge-requests/{id}/status": {
      "patch": {
        "summary": "Update merge request status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["open", "merged", "closed"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated merge request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeRequest"
                }
              }
            }
          }
        }
      }
    },
    "/api/stories": {
      "get": {
        "summary": "List stories",
        "parameters": [
          {
            "name": "mrId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of stories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserStory"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create story",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStory"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created story with analysis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "story": {
                      "$ref": "#/components/schemas/UserStory"
                    },
                    "analysis": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/stories/{id}/move": {
      "patch": {
        "summary": "Move story",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["index"],
                "properties": {
                  "parentId": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  },
                  "index": {
                    "type": "integer",
                    "minimum": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Moved story",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "story": {
                      "$ref": "#/components/schemas/UserStory"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tests": {
      "post": {
        "summary": "Create acceptance test",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptanceTest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created acceptance test",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptanceTest"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MergeRequest": {
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "status",
          "repository",
          "branch",
          "drifted",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["open", "merged", "closed"]
          },
          "repository": {
            "type": "string"
          },
          "branch": {
            "type": "string"
          },
          "drifted": {
            "type": "boolean"
          },
          "lastSyncAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserStory": {
        "type": "object",
        "required": [
          "id",
          "mrId",
          "title",
          "asA",
          "iWant",
          "soThat",
          "status",
          "order",
          "depth",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "mrId": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "asA": {
            "type": "string"
          },
          "iWant": {
            "type": "string"
          },
          "soThat": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["backlog", "in-progress", "done", "blocked"]
          },
          "estimateDays": {
            "type": "number",
            "nullable": true
          },
          "order": {
            "type": "integer"
          },
          "depth": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AcceptanceTest": {
        "type": "object",
        "required": [
          "id",
          "storyId",
          "title",
          "given",
          "when",
          "then",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "storyId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "given": {
            "type": "string"
          },
          "when": {
            "type": "string"
          },
          "then": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "passed", "failed"]
          },
          "notes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {}
        }
      }
    }
  }
}